requirements.txt
fastapi
uvicorn
pillow
matplotlib
numpy
tensorflow
ultralytics
